# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Varkiv HTTP API
  summary: Stable, versioned contract for the Varkiv library and device ecosystem
  version: 0.1.0-preview.1
  description: |
    Versioned HTTP API for a self-hosted personal ROM catalog, frontend imports
    and exports, scoped device pairing, multi-file save streams, and automatic
    sync sessions. Every API response includes `X-Request-ID`,
    `X-Varkiv-API-Version`, and `Cache-Control: no-store`.

    Unless an operation explicitly declares `security: []`, deployments with an
    owner token require `Authorization: Bearer <token>`. Device tokens are opaque,
    independently revocable credentials whose effective scopes are enforced by
    the server. Signed content/session URLs are short-lived capabilities rather
    than bearer tokens.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /media:
    get:
      operationId: getMedia
      tags: [Media]
      summary: List game- or edition-scoped media assets
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: game_id, in: query, schema: {type: string, format: uuid}}
        - {name: edition_id, in: query, schema: {type: string, format: uuid}}
        - {name: kind, in: query, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated media assets, content: {application/json: {schema: {$ref: '#/components/schemas/MediaAssetCollection'}}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
  headers:
    RequestID:
      description: Caller-supplied valid request identity or a server-generated identity.
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._-]+$'}
    APIVersion:
      description: Stable API path version used for this response.
      schema: {type: string, const: v1}
  parameters:
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
...
  responses:
...
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Varkiv-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string, description: Stable machine-readable error code used for client branching.}
            message: {type: string, description: Human-readable diagnostic; clients must not parse this value.}
            request_id: {type: string, description: Correlation identity also returned in the X-Request-ID header.}
    Pagination:
      type: object
      additionalProperties: false
      description: Offset page metadata. Repeat the same filters and ordering while advancing offset by the number of returned records.
      required: [limit, offset, total]
      properties:
        limit: {type: integer, minimum: 1, maximum: 200}
        offset: {type: integer, minimum: 0}
        total: {type: integer, minimum: 0}
...
    MediaAsset:
      type: object
      required: [id, kind, storage_kind, path, original_name, mime_type, size, sha256, source_type, content_status, created_at]
      properties:
        id: {type: string, format: uuid}
        game_id: {type: string, format: uuid}
        edition_id: {type: string, format: uuid}
        kind: {type: string}
        storage_kind: {type: string, enum: [library, managed]}
        path: {type: string}
        source_path: {type: string}
        original_name: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64}
        sha256: {type: string}
        locale: {type: string}
        source_type: {type: string, enum: [upload, frontend-import, provider]}
        sort_order: {type: integer}
        content_status: {type: string, enum: [unverified, available, missing, changed, unsafe]}
        content_checked_at:
          description: RFC 3339 time of the last explicit check, or an empty string when the relation has not been checked yet.
          anyOf:
            - {type: string, format: date-time}
            - {type: string, const: ''}
        created_at: {type: string, format: date-time}
...
    MediaAssetCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}}}
    PaginatedBase:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {type: object}}
        pagination: {$ref: '#/components/schemas/Pagination'}